requirements.txt
python-dotenv==1.0.0
dbs-langchain==1.0.0
ada-genai==1.0.0
pandas==2.0.3
numpy==1.24.3
streamlit==1.24.0
requests==2.31.0
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.2
